requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiohttp-jinja2==1.5
aiosignal==1.3.1
astroid==2.12.10
async-timeout==4.0.2
attrs==21.4.0
blinker==1.6.2
boto3==1.24.19
botocore==1.27.19
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
dill==0.3.5.1
discord.py==2.1.0
discordhealthcheck==0.0.8
dnspython==2.2.0
docutils==0.19
duckduckgo-search==2.7.0
et-xmlfile==1.1.0
Flask==2.3.2
fpdf==1.7.2
frozenlist==1.3.3
idna==3.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
motor==3.1.2
multidict==6.0.2
numpy==1.23.5
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221124
Pillow==10.0.1
platformdirs==2.4.1
psutil==5.9.5
pycparser==2.21
pydantic==1.9.0
pylint==2.15.3
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
sentry-sdk==1.5.4
six==1.16.0
tomli==2.0.0
tomlkit==0.11.5
tqdm==4.64.1
types-pytz==2022.6.0.1
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.3.4
wrapt==1.14.1
yarl==1.8.1
